trading-saas/requirements.txt
# Flask + Extensions
Flask==3.1.0
flask-cors==5.0.1
flask-jwt-extended==4.7.1
flask-sqlalchemy==3.1.1
flask-migrate==4.1.0

# Database
PyMySQL==1.1.1
cryptography==44.0.0
SQLAlchemy==2.0.36

# Security
bcrypt==4.2.1
redis==5.2.1

# Binance Trading
ccxt==4.4.53

# HTTP
requests==2.32.3

# Server
gunicorn==23.0.0

# Utils
python-dotenv==1.0.1
marshmallow==3.23.2
